refactor: landing page and dashboard files and folders structure
