--- src/front/js/component/Card/card.css
@@ -1,11 +1,0 @@
- 
-
-
-
-
-
-
-
-    
-    
-
