--- src/front/js/component/DashboardCard/dashboard-card.css
@@ -1,15 +1,0 @@
-.dashboard-card-container {
-    display: flex;
-    flex-flow: row nowrap;
-    justify-content: space-between;
-    gap: 3rem;
-}
-
-.dashboard-card {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    width: 100%;
-    border: 1px solid var(--dark-gray);
-    background: #fff;
-}
