--- src/front/js/component/Features/features.css
@@ -1,1 +1,0 @@
-
