--- src/front/js/component/Header/header.css
@@ -1,5 +1,0 @@
-
-
-
-
-
