+++ src/front/js/component/LandingPage/Slide/slide.css
@@ -1,0 +1,4 @@
+.carousel-item {
+    max-height: 250px;
+}
+
