--- src/front/js/component/Slide/slide.css
@@ -1,4 +1,0 @@
-.carousel-item {
-    max-height: 200px;
-}
-
