+++ src/front/js/index.js
@@ -2,9 +2,8 @@
 import React from "react";
 import ReactDOM from "react-dom";
 
-
 // Styles
-/* import "../assets/css/index.css"; */
+import "../assets/css/index.css";
 
 // Components
 import AppRouter from "./AppRouter.jsx";
