+++ src/front/js/layout/Layout.jsx
@@ -1,12 +1,20 @@
 import React from "react";
 import PropTypes from "prop-types";
 
+// Components
+// -- Landing Page
+import { Header } from "../component/LandingPage/Header/Header.jsx";
+import { Footer } from "../component/LandingPage/Footer/Footer.jsx";
+
 const Layout = (props) => {
+    console.log("propslayaout", props);
     return (
         <>
+            <Header />
             <main>
                 {props.children}
             </main>
+            <Footer />
         </>
     );
 };
